rename library folders
