+++ packages/library/src/hooks/usePhantomRedirectAdapter.ts
