+++ packages/library/src/index.ts
@@ -1,8 +1,8 @@
 export { default as PhantomRedirectAdapter } from './adapter'
 export * from './adapter'
 
-export { default as usePhantomRedirectProvider } from './hooks/usePhantomRedirectProvider'
-export * from './hooks/usePhantomRedirectProvider'
+export { default as usePhantomRedirectAdapter } from './hooks/usePhantomRedirectAdapter'
+export * from './hooks/usePhantomRedirectAdapter'
 
 export { default as shouldPhantomRedirect } from './utils/shouldPhantomRedirect'
 export * from './utils/shouldPhantomRedirect'
